requirements.txt
asgiref==3.3.1
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-environ==0.4.5
flake8==3.8.4
idna==2.10
mccabe==0.6.1
Pillow==8.1.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2021.1
requests==2.25.1
sorl-thumbnail==12.7.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
